+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## 0.0.1
+## [0.0.1] 2023-01-29
 
-* TODO: Describe initial release.
+* Initial release
+* Basic Animations: 
+- Fade
+- Slide
+- Scale
+
+* Route Aware Animation
+Combining Fade and Slide animations with navigation 
+
+* Route Aware Widget 
+Setup: Add routeObserver in material app (main)